tests,fixture: fix warnings with -Wsign-compare
  tests/test-fixture.c:130:28: error: comparison of integer expressions of different signedness: ‘gboolean’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
    130 |         if (modem_expected == n_modems) {
        |                            ^~

+++ plugins/tests/test-fixture.c
@@ -127,7 +127,7 @@ common_get_modem (TestFixture *fixture,
         n_modems = g_list_length (modems);
         g_assert_cmpuint (n_modems, <=, 1);
 
-        if (modem_expected == n_modems) {
+        if ((guint)modem_expected == n_modems) {
             if (modems) {
                 found = MM_OBJECT (g_object_ref (modems->data));
                 g_message ("Found modem at '%s'", mm_object_get_path (found));
